srv/receipts/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
psycopg2-binary==2.9.11
asyncpg==0.30.0

# Redis
redis==5.2.0

# Authentication
pyjwt==2.10.1
bcrypt==4.2.1
passlib==1.7.4

# AI/ML
aiohttp==3.11.10
pillow==11.0.0

# Excel Export
openpyxl==3.1.5

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
